flow: add egress VLAN membership verification
Reject flows when egress port is not a member of the VLAN.

+++ device.h
@@ -108,7 +108,7 @@ void device_set_bridge(struct device *dev, bool enabled);
 void device_vlan_add(struct device *dev, struct vlan *vlan);
 void device_vlan_remove(struct device *dev, int id);
 int device_vlan_get_input(struct device *dev, uint16_t xdp_vlan);
-uint16_t device_vlan_get_output(struct device *dev, int vid);
+int device_vlan_get_output(struct device *dev, int vid);
 bool device_vlan_state_forwarding(struct device *dev, int vid);
 void device_update(struct device *dev);
 void device_reset_offload_update(void);
